# PERF: Faster CategoricalIndex from categorical