@topper-123 2017-09-13
I am able to replicate this result (159 ms -> 9.88 µs).