@TomAugspurger 2017-09-14
Unrelated test failure https://travis-ci.org/pandas-dev/pandas/jobs/275540546#L1608 (already an issue at https://github.com/pandas-dev/pandas/issues/17510)